refactor(rebroadcast): extract reconstruction logic into db/queries/
Moves the inlined reconstructCampaignStatus + reconstructSnapshots
(~140 LOC of DISTINCT-ON + bucket-merge + wire-format assembly) out
of src/app/api/h1/rebroadcast/route.js into a new
src/db/queries/rebroadcast.mjs module. The route now imports both
functions and stays as pure POST orchestration.

No runtime behavior change — every Prisma call, $queryRaw shape,
return value, and null-handling branch is byte-for-byte preserved.
The route's existing test suite passes unchanged because tests mock
@/db/db directly, and the new query module imports @/db/db the same
way the route did.

Closes the rebroadcast-getcampaign-consolidation cluster (step 1)
from /desloppify issue #406.
